requirements.txt
# PenDoc - Penetration Testing Documentation Tool
# Python 3.10+

playwright>=1.40.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
jinja2>=3.1.0
colorama>=0.4.6
tqdm>=4.66.0
python-dateutil>=2.8.0

